src/server/requirements.txt
django==1.10.6
djangorestframework
